fix: Stars must be valid, cannot override values

+++ src/Modules/Tours/Explorer.Tours.API/Dtos/TourRatingDto.cs
@@ -13,7 +13,7 @@ namespace Explorer.Tours.API.Dtos
         public long TourExecutionId { get; set; }
         public int Stars { get; set; }
         public string Comment { get; set; }
-        public int CompletedProcentage { get; set; }
+        public double CompletedProcentage { get; set; }
         public int ThumbsUpCount { get; set; }
         public DateTime CreatedAt { get; set; }
     }
